Patch:Linux 3.4.11-3.4.105
Linux Upstream patches upto 3.4.105. Credits to original Maintainer Li Zefan and its former maintainer Greg Kroah-Hartman.

Highlights:
- Upstream patches
- Just reverted "futex: Unlock hb->lock in futex_wait_requeue_pi() error path"
- Ommitted some patches as they were already cherry-picked or manually added

+++ include/net/netfilter/nf_conntrack_extend.h
@@ -36,8 +36,8 @@ enum nf_ct_ext_id {
 
 struct nf_ct_ext {
 	struct rcu_head rcu;
-	u8 offset[NF_CT_EXT_NUM];
-	u8 len;
+	u16 offset[NF_CT_EXT_NUM];
+	u16 len;
 	char data[0];
 };
